requirements.txt
# requirements.txt

absl-py
accelerate
aiofiles
aiohttp
aiohttp-proxy
aiosignal
aniso8601
annotated-types
anyio
asgiref
asttokens
attrs
autogen
autogpt
autopep8
backoff
bcrypt
beautifulsoup4
blinker
blobfile
bs4
build
cachetools
certifi
cffi
charset-normalizer
chroma-hnswlib
chromadb
click
colorama
coloredlogs
comm
comtypes
contourpy
cssselect
cycler
dataclasses-json
dateparser
debugpy
decorator
deepdiff
Deprecated
dirtyjson
distlib
distro
dnspython
docopt
docutils
duckduckgo_search
email_validator
executing
fake-useragent
fastapi
fastapi-cli
fastjsonschema
feedfinder2
feedparser
ffmpy
filelock
fire
Flask
Flask-RESTful
flatbuffers
fonttools
frozendict
frozenlist
fsspec
ftfy
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
googleapis-common-protos
gradio
gradio_client
greenlet
grpcio
grpcio-status
h11
html2image
html2text
html5lib==1.1
httpcore
httplib2
httptools
httpx
huggingface-hub
humanfriendly
idna==3.6
ijson
importlib-metadata
importlib_resources
iniconfig
ipykernel
ipython
itsdangerous
jaraco.classes
jedi
jieba3k
Jinja2
jiter
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
keyboard
keyring
kiwisolver
kubernetes
langchain
langchain-community
langchain-core
langchain-ollama
langchain-text-splitters
langsmith
litellm
llama-cloud
llama-index
llama-index-agent-openai
llama-index-cli
llama-index-core
llama-index-embeddings-openai
llama-index-indices-managed-llama-cloud
llama-index-legacy
llama-index-llms-openai
llama-index-multi-modal-llms-openai
llama-index-program-openai
llama-index-question-gen-openai
llama-index-readers-file
llama-index-readers-llama-parse
llama-index-utils-openai
llama-parse
llama_models
llama_stack
llama_stack_client
lxml
lxml_html_clean
markdown-it-py
MarkupSafe
marshmallow
matplotlib
matplotlib-inline
mdurl
mesop
mmh3
monotonic==1.6
more-itertools
mpmath
msgpack
multidict
multitasking
mypy-extensions
nbformat
nest-asyncio
networkx
newspaper3k
nh3
nltk
numpy
oauthlib
ollama
onnxruntime
openai
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-asgi
opentelemetry-instrumentation-fastapi
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-util-http
ordered-set
orjson
outcome
overrides
packaging
pandas
parso
peewee
pillow
pipenv
pkginfo
platformdirs
playwright
pluggy
posthog
primp
prompt-toolkit
proto-plus
protobuf
psutil
pure-eval
py
pyasn1
pyasn1-modules
pycodestyle
pycparser
pycryptodomex
pydantic
pydantic-settings
pydantic_core
pydub
pyee
pygame
PyGetWindow
Pygments
PyMuPDF
PyMuPDFb
pyparsing
pypdf
pyperclip
PyPika
pypiwin32
pyproject_hooks
pyreadline3
PyRect
PySocks
pytesseract
pytest
python-dateutil
python-dotenv
python-multipart
pyttsx3
pytz
pywin32
pywin32-ctypes
PyYAML
pyzmq
readme-renderer
referencing
regex
requests
requests-file
requests-oauthlib
requests-toolbelt
retry
rfc3986
rich
rpds-py
rsa==4.9
ruff
safetensors
selenium
selenium-stealth
semantic-version
setuptools
sgmllib3k
shellingham
simplejson
six
sniffio
sortedcontainers
soupsieve==2.5
SQLAlchemy
stack-data
starlette
striprtf
sympy
tabulate
tenacity
termcolor
tiktoken
tinysegmenter==0.3
tk
tldextract
tokenizers
tokentrim
toml
tomlkit
torch
tornado==6.4
tqdm
traitlets
transformers
trio
trio-websocket
twine
typer
typing-inspect
typing_extensions
tzdata
tzlocal==5.2
ujson
uritemplate
urllib3
uvicorn
virtualenv
watchdog
watchfiles
wcwidth
webdriver-manager
webencodings
websocket-client
websockets
Werkzeug
whisper
win10toast==0.9
wrapt
wsproto
yarl
yaspin
yfinance
youtube-comment-downloader
zipp
